draw_digits/pyproject.toml
[project]
name = "drawing-window-gtk"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.1",
    "pycairo>=1.29.0",
    "pygobject>=3.54.5",
    "pygobject-stubs>=2.16.0",
    "pytest>=9.0.2",
]
